app/domain-mon/requirements.txt
requests>=2.31.0
python-whois>=0.9.4
python-dateutil>=2.8.2
